refactor(init): split run-device command handler
