+++ src/index.ts
@@ -29,12 +29,13 @@ import { setChannel } from './channel/set'
 import { generateDocs } from './docs'
 import { defaultStarRepo } from './github'
 import { starAllRepositoriesCommand, starRepositoryCommand } from './github-command'
-import { initApp, testRunDeviceCommand } from './init'
+import { initApp } from './init'
 import { createKey, deleteOldKey, saveKeyCommand } from './key'
 import { login } from './login'
 import { startMcpServer } from './mcp/server'
 import { addOrganization, deleteOrganization, listMembers, listOrganizations, setOrganization } from './organization'
 import { probe } from './probe'
+import { testRunDeviceCommand } from './run-device'
 import { getUserId } from './user/account'
 import { formatError } from './utils'
